@@ -1,28 +1,30 @@
 <?php
 
+use TYPO3\CMS\Core\Imaging\IconProvider\SvgIconProvider;
+
 return [
     'apps-pagetree-folder-cartproducts-products' => [
-        'provider' => \TYPO3\CMS\Core\Imaging\IconProvider\SvgIconProvider::class,
+        'provider' => SvgIconProvider::class,
         'source' => 'EXT:cart_products/Resources/Public/Icons/apps_pagetree_folder_cartproducts_products.svg',
     ],
     'apps-pagetree-page-cartproducts-products' => [
-        'provider' => \TYPO3\CMS\Core\Imaging\IconProvider\SvgIconProvider::class,
+        'provider' => SvgIconProvider::class,
         'source' => 'EXT:cart_products/Resources/Public/Icons/apps_pagetree_page_cartproducts_products.svg',
     ],
     'ext-cartproducts-extension-icon' => [
-        'provider' => \TYPO3\CMS\Core\Imaging\IconProvider\SvgIconProvider::class,
+        'provider' => SvgIconProvider::class,
         'source' => 'EXT:cart_products/Resources/Public/Icons/Extension.svg',
     ],
     'ext-cartproducts-wizard-icon-list' => [
-        'provider' => \TYPO3\CMS\Core\Imaging\IconProvider\SvgIconProvider::class,
+        'provider' => SvgIconProvider::class,
         'source' => 'EXT:cart_products/Resources/Public/Icons/cartproducts_plugin_wizard-list.svg',
     ],
     'ext-cartproducts-wizard-icon-show' => [
-        'provider' => \TYPO3\CMS\Core\Imaging\IconProvider\SvgIconProvider::class,
+        'provider' => SvgIconProvider::class,
         'source' => 'EXT:cart_products/Resources/Public/Icons/cartproducts_plugin_wizard-show.svg',
     ],
     'ext-cartproducts-wizard-icon-teaser' => [
-        'provider' => \TYPO3\CMS\Core\Imaging\IconProvider\SvgIconProvider::class,
+        'provider' => SvgIconProvider::class,
         'source' => 'EXT:cart_products/Resources/Public/Icons/cartproducts_plugin_wizard-teaser.svg',
     ],
 ];
